Addition Of Ability To Add Or Remove Data
Added the ability to add or remove data by the ability of viewing nodes and member data through a drop-down list. In the current version aesthetics and consistencies do not exist in a future patch, additionally menus for other forms will include this feature as well as a more consistent, better aesthetically pleasing menu.